Install React router and re order folder structure
