+++ web/package.json
@@ -22,9 +22,11 @@
     "@types/react-dom": "^16.9.0",
     "@types/react-redux": "^7.1.7",
     "nhsuk-react-components": "^1.2.8",
+    "normalize.css": "^8.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "typescript": "~3.8.2"
   },
@@ -35,10 +37,14 @@
   "name": "web",
   "private": false,
   "scripts": {
-    "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "postbuild": "react-snap",
+    "start": "react-scripts start",
+    "test": "react-scripts test"
   },
-  "version": "0.1.0"
+  "version": "0.1.0",
+  "devDependencies": {
+    "@types/react-router-dom": "^5.1.7"
+  }
 }
