+++ web/src/app/store.ts
@@ -1,5 +1,5 @@
 import { configureStore, ThunkAction, Action } from '@reduxjs/toolkit';
-import counterReducer from '../features/counter/counterSlice';
+import counterReducer from './features/counter/counterSlice';
 
 export const store = configureStore({
   reducer: {
