+++ web/src/pages/404.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const ErrorPage: React.FC = () => <h1>ErrorPage</h1>;
+
+export default ErrorPage;
