+++ web/src/pages/home.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const Home: React.FC = () => <h1>Home</h1>;
+
+export default Home;
